Refactor trigger window into separate files
