+++ crates/gwaihir-client/src/app.rs
@@ -10,13 +10,12 @@ use crate::{
         outputs::{sensor_output::SensorOutput, sensor_outputs::SensorOutputs},
     },
     tray_icon::{hide_to_tray, TrayIconData},
-    triggers::{BehaviorOnTrigger, TriggerManager, Update},
+    triggers::{ui::TriggersWindow, BehaviorOnTrigger, TriggerManager, Update},
     ui::{
         add_fake_user_window::AddFakeUserWindow,
         network_window::NetworkWindow,
         raw_data_window::{RawDataWindow, TimestampedData},
         time_formatting::nicely_formatted_datetime,
-        triggers_window::TriggersWindow,
         ui_extension_methods::UIExtensionMethods,
         widgets::auto_launch_checkbox::AutoLaunchCheckboxUiExtension,
     },
