+++ crates/gwaihir-client/src/ui/mod.rs
@@ -2,6 +2,5 @@ pub mod add_fake_user_window;
 pub mod network_window;
 pub mod raw_data_window;
 pub mod time_formatting;
-pub mod triggers_window;
 pub mod ui_extension_methods;
 pub mod widgets;
